// src/app/models/PokemonDetails.ts
export interface PokemonDetails {
  base_happiness: number;
  capture_rate: number;
  color: { name: string; url: string };
  egg_groups: { name: string; url: string }[];
  evolution_chain: { url: string };
  evolves_from_species: { name: string; url: string };
  flavor_text_entries: { flavor_text: string, language: { name: string; url: string }; version: { name: string; url: string } }[];
  form_descriptions: unknown[];
  forms_switchable: boolean;
  gender_rate: number;
  genera: { genus: string; language: { name: string; url: string } }[];
  generation: { name: string; url: string };
  growth_rate: { name: string; url: string };
  habitat: { name: string; url: string };
  has_gender_differences: boolean;
  hatch_counter: number;
  id: number;
  is_baby: boolean;
  is_legendary: boolean;
  is_mythical: boolean;
  name: string;
  names: {language: { name: string, url: string }; name: string }[];
  order: number;
  pal_park_encounters: { area: { name: string; url: string }; base_score: number; rare: number }[];
  pokedex_numbers: {entry_number: number; pokedex: { name: string; url: string }}[];
  shape: { name: string; url: string };
  varieties: { isDefault: boolean; pokemon: { name: string; url: string } }[];
}